Update connection in tests
